+++ vidya/__init__.py
@@ -2,5 +2,5 @@ import pathlib
 
 ROOT = pathlib.Path(__file__).parent.parent
 
-__version__ = '0.1.6'
-__update_date__ = '2025-06-26_19:47:42'
+__version__ = '0.1.7'
+__update_date__ = '2025-06-30_15:08:30'